Documentation

start and start!

AsynchronousIterativeAlgorithms.start — Function

start(algorithm::AbstractAlgorithm{Q,A}, problem_constructor::Function, stopat::Union{Tuple{Int64, Int64, Float64}, Tuple{Int64, Int64, Float64, Float64}}; saveat=(0,0), save_answers=false, pids=workers(), synchronous=false, distance::Function=(x::Q,y::Q)->norm(x-y), resilience=0, verbose=1) where {Q,A}

Solve the distributed problem returned by problem_constructor using the algorithm.

Arguments

algorithm::AbstractAlgorithm{Q,A}: subtyping AbstractAlgorithm{Q,A} and implementing its functor calls

problem_constructor::Function: this function should return the process pid's problem when it calls problem_constructor(pid::Int64) (for any remote pids and on the current pid)

stopat::Union{Tuple{Int64, Int64, Float64}, Tuple{Int64, Int64, Float64, Float64}}: (i, e, t) or (i, e, t, p)

i: maximum number of iterations

e: maximum number of epochs (all workers have answered at least e times)

t: maximum starttime t (in seconds)

p: required precision (in terms of distance between the last two queries)

Keywords

saveat=(0,0)::Tuple{Int64, Int64}: query iterates (::Q) sent by the central nodes are recorded every i > 0 iterations, e > 0 epochs in saveat=(i, e)

save_answers=false::Bool: answer iterates (::A) sent by the workers are recorded

pids::Vector{Int64}=workers(): pids of the active workers, you can start a non-distributed (and necessarily synchronous) version of your algorithm with pids=[1]

synchronous=false: if synchronous=true, the central node waits for all workers to answer before making a step

distance::Function=(x::Q,y::Q)->norm(x-y): function used to compute the distance between the last two queries

resilience::Int64=0: number of workers allowed to fail before the execution is stopped

verbose=1: if > 0, a progress bar is displayed

Returns

NamedTuple: a record of the queries and the iterations, epochs, timestamps at which they were recorded, as well as answer_count of each worker (if save_answers is true, the answers will be recorded with their worker provenance in answer_origin)

Throws

ArgumentError: if the arguments don't match the specifications.

start(algorithm::AbstractAlgorithm{Q,A}, distributedproblem::DistributedObject{M}, stopat::Union{Tuple{Int64, Int64, Float64}, Tuple{Int64, Int64, Float64, Float64}}; saveat=(0,0), saveanswers=false, pids=workers(), synchronous=false, distance::Function=(x::Q,y::Q)->norm(x-y), resilience=0, verbose=1) where {Q,A,M}

Solve the distributed_problem using the algorithm. Similar to the original start function but instead of a problem_constructor::Function, a distributed_problem::DistributedObject should be passed. distributed_problem should reference a problem on the remote pids and on the current pid.

AsynchronousIterativeAlgorithms.start! — Function

start!(algorithm::AbstractAlgorithm{Q,A}, problem_constructor::Function, stopat::Union{Tuple{Int64, Int64, Float64}, Tuple{Int64, Int64, Float64, Float64}}; saveat=(0,0), save_answers=false, pids=workers(), synchronous=false, distance::Function=(x::Q,y::Q)->norm(x-y), resilience=0, verbose=1) where {Q,A}

Same as start but start! uses a deep copy of your algorithm and won't modify it. This version enables modifications. This can be useful to record information during the execution for example.

start!(algorithm::AbstractAlgorithm{Q,A}, distributed_problem::DistributedObject{M}, stopat::Union{Tuple{Int64, Int64, Float64}, Tuple{Int64, Int64, Float64, Float64}}; saveat=(0,0), save_answers=false, pids=workers(), synchronous=false, distance::Function=(x::Q,y::Q)->norm(x-y), resilience=0, verbose=1) where {Q,A,M}

Same as start but start! uses a deep copy of your algorithm and won't modify it. This version enables modifications. This can be useful to record information during the execution for example.

Algorithm

AsynchronousIterativeAlgorithms.AbstractAlgorithm — Type

AbstractAlgorithm{Q,A}

Types subtyping AbstractAlgorithm should be callable with the following signatures:

(algorithm::AbstractAlgorithm{Q,A})(problem::Any) where {Q,A} the initialization step that create the first query iterate q::Q

(algorithm::AbstractAlgorithm{Q,A})(q::Q, problem::Any) where {Q,A} is the step perfromed by the wokers when they receive a query q::Q from the central node

(algorithm::AbstractAlgorithm{Q,A})(a::A, worker::Int64, problem::Any) where {Q,A} is the step performed by the central node when receiving an answer a::A from a worker

when start takes the keyword synchronous=true, (algorithm::AbstractAlgorithm{Q,A})(as::Vector{A}, workers::Vector{Int64}, problem::Any) where {Q,A} is the step performed by the central node when receiving the answers as::Vector{A} from all the workers in pids

They can additionally define:

AsynchronousIterativeAlgorithms.stopnow(algorithm::MyAlgorithm) (with the trait AsynchronousIterativeAlgorithms.Stoppability(::MyAlgorithm) = Stoppable()) to add a stopping condition to start's (iterations, epochs, time) stopping condition

Algorithm templates

AsynchronousIterativeAlgorithms.AggregationAlgorithm — Type

AggregationAlgorithm{Q,A}(initialize::Function, aggregate::Function, query::Function, answer::Function, initial_answer::A; pids=workers()) where {Q,A}

Distributed algorithm that writes: q_j <- query(aggregate([answer(q_i) for i in connected])) Where a "connected" worker is a worker that has answered at least once.

The function parameters should have the following signature

initialize(problem::Any))

aggregate(a::Vector{A}, workers::Vector{Int64}) where workers lists the provenance of the elements of a

query(a::A, problem::Any)

answer(q::Q, problem::Any)

(Not memory optimized: length(pids) answers are stored on the central worker at all times)

AsynchronousIterativeAlgorithms.AveragingAlgorithm — Type

AveragingAlgorithm{Q,A}(initialize::Function,m= query::Function, answer::Function; pids=workers(), weights=ones(length(pids))) where {Q,A}

Distributed algorithm that writes: q_j <- query(weighted_average([answer(q_i) for i in connected])) Where a "connected" worker is a worker that has answered at least once.

The function parameters should have the following signature

initialize(problem::Any))

query(a::A, problem::Any)

answer(q::Q, problem::Any)

(Memory optimized: only the equivalent of one answer is stored on the central worker at all times)