Documentation

start and start!

AsynchronousIterativeAlgorithms.start — Function

start(algorithm, problem_constructor, stopat; kwargs...)
start(algorithm, distributed_problem, stopat; kwargs...)

Solve the distributed problem returned by problem_constructor (or referenced by distributed_problem) using the algorithm until the stopat conditions are reached.

Arguments

algorithm::AbstractAlgorithm{Q,A}: subtypes AbstractAlgorithm{Q,A} and implementing its functor calls

problem_constructor::Function: for each pid in {pids ⋃ current pid}, process pid calling problem_constructor(pid) should return the process' assigned problem

distributed_problem::DistributedObject: for each pid in {pids ⋃ current pid}, distributed_problem should reference process pid's assigned problem on pid

stopat::Tuple: (i, e, t) or (i, e, t, p)

i::Int64: maximum number of iterations

e::Int64: maximum number of epochs (all workers have answered at least e times)

t::Float64: maximum starttime t (in seconds)

p::Float64: required precision (in terms of distance between the last two queries)

Keywords

saveat=(0,0)::Tuple{Int64, Int64}: query iterates (::Q) sent by the central nodes are recorded every i > 0 iterations, e > 0 epochs in saveat=(i, e)

save_answers=false::Bool: answer iterates (::A) sent by the workers are recorded

pids::Vector{Int64}=workers(): pids of the active workers, you can start a non-distributed (and necessarily synchronous) version of your algorithm with pids=[1]

synchronous::Bool=false: if synchronous=true, the central node waits for all workers to answer before making a step

distance::Function=(x::Q,y::Q)->norm(x-y): function used to compute the distance between the last two queries

resilience::Int64=0: number of workers allowed to fail before the execution is stopped

verbose::Int64=1: if > 0, a progress bar is displayed

Returns

NamedTuple: a record of the queries and the iterations, epochs, timestamps at which they were recorded, as well as answer_count of each worker, additionally,

if save_answers=true, the provenance of the answers will be recorded in answers_origin

if p is set in stopat, the distance between query iterates will be recorded in precisions (note that the first precision will always be Inf at the first iteration)

Throws

ArgumentError: if the arguments don't match the specifications.

AsynchronousIterativeAlgorithms.start! — Function

start!(algorithm, problem_constructor, stopat; kwargs...)
start!(algorithm, distributed_problem, stopat; kwargs...)

Same args and kwargs as start but start! uses a deep copy of your algorithm and won't modify it. start!, on the other hand, enables modifications of your algorithm, which can be useful to record information during the execution for example.

AbstractAlgorithm

The algorithm you pass to start should subtype AbstractAlgorithm{Q,A}.

AsynchronousIterativeAlgorithms.AbstractAlgorithm — Type

AbstractAlgorithm{Q,A}

Types subtyping AbstractAlgorithm should be callable with the following signatures:

(algorithm::AbstractAlgorithm{Q,A})(problem::Any) where {Q,A} the initialization step that create the first query iterate q::Q

(algorithm::AbstractAlgorithm{Q,A})(q::Q, problem::Any) where {Q,A} is the step perfromed by the wokers when they receive a query q::Q from the central node

(algorithm::AbstractAlgorithm{Q,A})(a::A, worker::Int64, problem::Any) where {Q,A} is the step performed by the central node when receiving an answer a::A from a worker

when start takes the keyword synchronous=true, (algorithm::AbstractAlgorithm{Q,A})(as::Vector{A}, workers::Vector{Int64}, problem::Any) where {Q,A} is the step performed by the central node when receiving the answers as::Vector{A} from all the workers in pids

They can additionally define:

AsynchronousIterativeAlgorithms.stopnow(algorithm::MyAlgorithm) (with the trait AsynchronousIterativeAlgorithms.Stoppability(::MyAlgorithm) = Stoppable()) to add a stopping condition to start's (iterations, epochs, time) stopping condition

Algorithm templates

The two following algorithms already subtype AbstractAlgorithm{Q,A} and are ready to use in start.

AsynchronousIterativeAlgorithms.AggregationAlgorithm — Type

AggregationAlgorithm{Q,A}(args...; kwargs...) -> AggregationAlgorithm{Q,A} <: AbstractAlgorithm{Q,A}

Distributed algorithm that writes: q_j <- query(aggregate([answer(q_i) for i in connected])) Where a "connected" worker is a worker that has answered at least once. (Not memory optimized: length(pids) answers are stored on the central worker at all times)

Arguments

initialize::Function: with signature initialize(problem::Any)

aggregate::Function: with signature aggregate(a::Vector{A}, workers::Vector{Int64}) where workers lists the provenance of the elements of a

query::Function: with signature query(a::A, problem::Any)

answer::Function: with signature answer(q::Q, problem::Any)

Keyword

pids=workers():

AsynchronousIterativeAlgorithms.AveragingAlgorithm — Type

AveragingAlgorithm{Q,A}(args...; kwargs...) -> AveragingAlgorithm{Q,A} <: AbstractAlgorithm{Q,A}

Distributed algorithm that writes: q_j <- query(weighted_average([answer(q_i) for i in connected])) Where a "connected" worker is a worker that has answered at least once. (Memory optimized: only the equivalent of one answer is stored on the central worker at all times)

Arguments

initialize::Function: with signature initialize(problem::Any)

query::Function: with signature query(a::A, problem::Any)

answer::Function: with signature answer(q::Q, problem::Any)

Keywords

pids::Vector{Int64}=workers(): pids of the active workers

weights::Vector{Float64}=ones(length(pids)): weight of each pid in the weighted average